requirements.txt
Pillow
./games